refactor: create /hooks/post directory
