+++ src/hooks/post/useInformation.ts
